refactor: collapse guard module into hl7v2
