+++ crates/hl7v2-guard/Cargo.toml
@@ -12,5 +12,4 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-serde = { workspace = true, features = ["derive"] }
-hl7v2-core = { workspace = true }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = ["experimental-guard"] }
