+++ crates/hl7v2/Cargo.toml
@@ -42,7 +42,7 @@ synthetic = [
 ]
 redact = []
 lifecycle = ["dep:hl7v2-lifecycle"]
-experimental-guard = ["dep:hl7v2-guard"]
+experimental-guard = ["dep:serde"]
 
 [dependencies]
 hl7v2-model = { version = "1.2.0", path = "../hl7v2-model" }
@@ -52,6 +52,7 @@ hl7v2-query = { version = "1.2.0", path = "../hl7v2-query" }
 hl7v2-parser = { version = "1.2.0", path = "../hl7v2-parser" }
 hl7v2-writer = { version = "1.2.0", path = "../hl7v2-writer" }
 thiserror = { workspace = true }
+serde = { workspace = true, features = ["derive"], optional = true }
 
 hl7v2-json = { version = "1.2.0", path = "../hl7v2-json", optional = true }
 hl7v2-normalize = { version = "1.2.0", path = "../hl7v2-normalize", optional = true }
@@ -69,7 +70,6 @@ hl7v2-faker = { version = "1.2.0", path = "../hl7v2-faker", optional = true }
 hl7v2-corpus = { version = "1.2.0", path = "../hl7v2-corpus", optional = true }
 hl7v2-gen = { version = "1.2.0", path = "../hl7v2-gen", optional = true }
 hl7v2-lifecycle = { version = "1.2.0", path = "../hl7v2-lifecycle", optional = true }
-hl7v2-guard = { version = "1.2.0", path = "../hl7v2-guard", optional = true }
 
 [lints]
 workspace = true
