+++ crates/hl7v2/src/experimental.rs
@@ -1,0 +1,4 @@
+//! Experimental APIs.
+
+#[cfg(feature = "experimental-guard")]
+pub mod guard;
